% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rf_parallel.R
\name{cv_rf_parallel}
\alias{cv_rf_parallel}
\title{cv_rf_parallel
functional to determine the optimal number of variables used in each tree of the RF model using parallel computing}
\usage{
cv_rf_parallel(
  rf_dt_temp,
  K = 10,
  mtry_range,
  num_cores = 4,
  sample_ratio = 0.5,
  seeds = 123
)
}
\arguments{
\item{rf_dt_temp}{data.frame with first column as response variable.}

\item{K}{K-fold cross validation}

\item{mtry_range}{range of number of variables used in each trees. example:c(1:10)}

\item{num_cores}{how many cores are used in parallel computing. default as 10}

\item{sample_ratio}{in case the data was too big, part of the data is used in the cross validation to enhance the efficiency of the algorithm; default as 0.5}
}
\value{
data.frame with 2 columns indicating the cross validation error of each mtry
}
\description{
cv_rf_parallel
functional to determine the optimal number of variables used in each tree of the RF model using parallel computing
}
