% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shap.R
\name{plot_shap}
\alias{plot_shap}
\title{plot_shap
variable importance plot and variable interaction plot}
\usage{
plot_shap(
  plot_data,
  new_data,
  shap_df,
  variable_list = NULL,
  id_list = NULL,
  interaction_variable_list = NULL
)
}
\arguments{
\item{plot_data}{plot data from SHAP_analysis}

\item{new_data}{data need SHAP analysis. Response variables shouldn't be included.}

\item{shap_df}{shap values from SHAP_analysis. A character vector}

\item{variable_list}{variable list for importance plotting. A character vector. if null all features are plotted}

\item{id_list}{observation list for plotting. A character vector. if null, all observations are plotted.}

\item{interaction_variable_list}{variable list for interaction plotting. A character vector. odd elements and even elements are effect1 and effect2 respectively. If null, plot won't be produced.}
}
\value{
ggplot objects
}
\description{
plot_shap
variable importance plot and variable interaction plot
}
