% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shap_analysis.R
\name{shap_analysis}
\alias{shap_analysis}
\title{shap_analysis
function to carry out SHAP analysis using fastshap}
\usage{
shap_analysis(model, new_data, seed = 123, simulation_times = 10)
}
\arguments{
\item{model}{best model produced from CV}

\item{new_data}{data need SHAP analysis. Response variables shouldn't be included.}

\item{seed}{random seed. default as 123}

\item{simulation_times}{simulation times}
}
\value{
SHAP res from fastshap
}
\description{
shap_analysis
function to carry out SHAP analysis using fastshap
}
