% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAPonRF.R
\name{SHAPonRF}
\alias{SHAPonRF}
\title{SHAPonRF
full process of SHAP analysis on RandomForest model}
\usage{
SHAPonRF(
  rf_dt,
  mtry_range = 2:ncol(),
  K_fold_CV = 10,
  num_cores = 4,
  sample_ratio = 1,
  need_scale_shap = T,
  need_odd_remove = T,
  variable_list = NULL,
  id_list = NULL,
  inter_list = NULL,
  seed = 123
)
}
\arguments{
\item{rf_dt}{data.frame with first column indicating the response variable}

\item{mtry_range}{range of number of variables used in each trees. example:c(1:10)}

\item{K_fold_CV}{K-fold cross validation}

\item{num_cores}{how many cores are used in parallel computing. default as 10}

\item{sample_ratio}{in case the data was too big, part of the data is used in the cross validation to enhance the efficiency of the algorithm; default as 0.5}

\item{need_scale_shap}{SHAP value scale for visualization}

\item{need_odd_remove}{delete odd value SHAP value scale for visualization}

\item{variable_list}{variable list for importance plotting. A character vector. if null all features are plotted}

\item{id_list}{observation list for plotting. A character vector. if null, all observations are plotted.}

\item{inter_list}{variable list for interaction plotting. A character vector. odd elements and even elements are effect1 and effect2 respectively. If null, plot won't be produced.}

\item{seed}{random seed;default 123}
}
\value{
list(cv_mse:cross validation error ; best_rf: best randomforest model,shap_res: results from SHAP; SHAP_plots:visualization of SHAP analysis
}
\description{
SHAPonRF
full process of SHAP analysis on RandomForest model
}
